Rename uvicorn to run
